[DF-301] Refactoring classes paths (#14)
