+++ hive_metastore_client/hive_mestastore_client.py
@@ -1,5 +1,5 @@
 """Hive Metastore Client main class."""
-from hive_metastore_client.hive_mestastore_client_connector import (
+from hive_metastore_client.clients.hive_mestastore_client_connector import (
     HiveMetastoreClientConnector,
 )
