--- hive_metastore_client/services/__init__.py
@@ -1,1 +1,0 @@
-"""Service Classes."""
