+++ thrift_files/libraries/thrift_hive_metastore_client/ThriftHiveMetastore.py
@@ -9,22 +9,13 @@
 #
 
 import logging
-import sys
 
-from thrift.protocol.TProtocol import TProtocolException
 from thrift.Thrift import (
-    TApplicationException,
-    TException,
-    TFrozenDict,
-    TMessageType,
     TProcessor,
-    TType,
 )
-from thrift.transport import TTransport
-from thrift.TRecursive import fix_spec
 
-from hive_metastore_client.clients.thrift_hive_metastore_client.ttypes import *
-from hive_metastore_client.services.thrift_facebook_service import FacebookService
+from thrift_files.libraries.thrift_facebook_service import FacebookService
+from thrift_files.libraries.thrift_hive_metastore_client.ttypes import *
 
 all_structs = []
