+++ thrift_files/libraries/thrift_hive_metastore_client/constants.py
